Fixed REALLY stupid issue that was breaking lighting.

+++ FinModelUtility/Fin/Fin/src/shaders/glsl/ColorShaderSourceGlsl.cs
@@ -29,7 +29,7 @@ namespace fin.shaders.glsl {
 
            uniform vec4 diffuseColor;
            uniform float {GlslConstants.UNIFORM_SHININESS_NAME};
-           uniform float {GlslConstants.UNIFORM_USE_LIGHTING_NAME};
+           uniform int {GlslConstants.UNIFORM_USE_LIGHTING_NAME};
 
            out vec4 fragColor;
